+++ dirti.aldyn/S1/outputList.hpp
@@ -3,10 +3,11 @@
 #include <ostream>
 #include <list>
 #include <utility>
+#include <stdexcept>
 
 namespace dirti
 {
-  void outputList(const std::list< std::pair< std::string, std::list< int >>> mainList, std::ostream& out);
+  void outputList(const std::list< std::pair< std::string, std::list< size_t >>> mainList, std::ostream& out);
 }
 
 #endif
